requirements.txt
aiogram            2.23.1
aiohttp            3.8.3
aiosignal          1.3.1
async-timeout      4.0.2
attrs              22.1.0
Babel              2.9.1
certifi            2022.9.24
charset-normalizer 2.1.1
frozenlist         1.3.3
greenlet           2.0.1
idna               3.4
magic-filter       1.0.9
multidict          6.0.3
pip                22.3.1
python-dotenv      0.21.0
pytz               2022.6
setuptools         65.5.1
SQLAlchemy         1.4.44
wheel              0.38.4
yarl               1.8.2

